+++ soxai_data/__init__.py
@@ -1,5 +1,5 @@
 """SOXAI Data Science Library."""
 
-from .soxai_data import DataLoader
+from .soxai_data import DataLoader, MAX_RANGE_DAYS
 
-__all__ = ['DataLoader']
+__all__ = ['DataLoader', 'MAX_RANGE_DAYS']